Rename: url이동에 따른 파일 수정 및 불필요한 파일 삭제
