--- front/pages/intro/IntroStyle.tsx
@@ -1,1 +1,0 @@
-import styled, { keyframes } from 'styled-components';
